Disable pinch zoom and other multi finger events
Best would be if we could just disable pinch zoom, but didn't find a way
to do that, so for now we disable all multi finger events. Mac users
that are used to scrolling with two fingers might be confused until they
learn that scrolling is done with one finger.